Fixed footer position & added Carousel
Footer will now be pushed to the bottom of the page and not cover up
content. Added a carousel to the page. However, the glyphicons do not
work with (my) firefox, I’m not sure about others. but they work with
google chrome and safari so that’s something we need to figure out